api-server-django-main/requirements.txt
Django==3.2.13
djangorestframework==3.13.1
PyJWT==2.4.0
django-cors-headers==3.13.0
gunicorn==20.1.0
django-environ==0.8.1
google-ads==18.0.0
google-api-core==2.8.2
google-auth==2.10.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
facebook-business==14.0.0
facebook-sdk==3.1.0
